Moved CxbxXbdm + renamed Common -> common
