+++ src/common/CxbxCommon.cpp
