+++ src/common/CxbxCommon.h
