+++ src/common/CxbxDebugger.h
