+++ src/common/CxbxUtil.h
