+++ src/common/EmuEEPROM.cpp
