+++ src/common/EmuEEPROM.h
