+++ src/common/Logging.cpp
