+++ src/common/Settings.cpp
