+++ src/common/Win32/EmuShared.cpp
