+++ src/common/Win32/InlineFunc.cpp
