+++ src/common/Xbe.cpp
