+++ src/common/XbePrinter.cpp
