+++ src/common/XbePrinter.h
