+++ src/common/xbdm/CxbxXbdm.cpp
